+++ language_server/src/dap/dap_feature.h
@@ -77,6 +77,8 @@ private:
     void stopped(hlasm_plugin::parser_library::sequence<char> reason,
         hlasm_plugin::parser_library::sequence<char> addtl_info) override;
     void exited(int exit_code) override;
+    void mnote(unsigned char level, hlasm_plugin::parser_library::sequence<char> text) override;
+    void punch(hlasm_plugin::parser_library::sequence<char> text) override;
 
     parser_library::debugger_configuration_provider& dc_provider;
     std::optional<hlasm_plugin::parser_library::debugging::debugger> debugger;