+++ parser_library/src/processing/statement_processors/ordinary_processor.cpp
@@ -40,14 +40,15 @@ ordinary_processor::ordinary_processor(const analyzing_context& ctx,
     processing_state_listener& state_listener,
     statement_fields_parser& parser,
     opencode_provider& open_code,
-    processing_manager& proc_mgr)
+    processing_manager& proc_mgr,
+    output_handler* output)
     : statement_processor(processing_kind::ORDINARY, ctx)
     , branch_provider_(branch_provider)
     , lib_info(lib_provider)
     , eval_ctx { *ctx.hlasm_ctx, lib_info, *this }
     , ca_proc_(ctx, branch_provider, lib_provider, state_listener, open_code)
     , mac_proc_(ctx, branch_provider, lib_provider)
-    , asm_proc_(ctx, branch_provider, lib_provider, parser, open_code, proc_mgr)
+    , asm_proc_(ctx, branch_provider, lib_provider, parser, open_code, proc_mgr, output)
     , mach_proc_(ctx, branch_provider, lib_provider, parser, proc_mgr)
     , finished_flag_(false)
     , listener_(state_listener)